LayoutTests/fast/history/page-cache-MessagePort-pending-message-expected.txt

Tests that a page that has a MessagePort with a pending message can enter PageCache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS e.data is "foo"
PASS afterPageCacheRestore is true
PASS successfullyParsed is true

TEST COMPLETE
